+++ Ui/Admin/Action/PortalNodeExtensionActivateUi.php
@@ -14,9 +14,8 @@ use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalExtension\PortalExt
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNode\PortalNodeGetActionInterface;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionActivate\PortalNodeExtensionActivatePayload;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\PortalNode\PortalNodeExtensionActivateUiActionInterface;
-use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionDoesNotSupportPortalNodeException;
-use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionIsAlreadyActiveOnPortalNodeException;
-use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionMissingException;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\NoMatchForPackageQueryException;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionsAreAlreadyActiveOnPortalNodeException;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalNodeMissingException;
 
 final class PortalNodeExtensionActivateUi implements PortalNodeExtensionActivateUiActionInterface
@@ -47,38 +46,39 @@ final class PortalNodeExtensionActivateUi implements PortalNodeExtensionActivate
 
     public function activate(PortalNodeExtensionActivatePayload $payload): void
     {
+        if ($payload->getPortalExtensionQueries() === []) {
+            return;
+        }
+
         $portalNodeKey = $payload->getPortalNodeKey();
         $portalNodeGetCriteria = new PortalNodeGetCriteria(new PortalNodeKeyCollection([$portalNodeKey]));
 
         foreach ($this->portalNodeGetAction->get($portalNodeGetCriteria) as $portalNodeGetResult) {
             $portalExtensionActivePayload = new PortalExtensionActivatePayload($portalNodeKey);
-            $portalExtensions = $this->portalLoader->getPortalExtensions();
+            $portalExtensions = $this->portalLoader->getPortalExtensions()->bySupport($portalNodeGetResult->getPortalClass());
             $portalExtensionState = $this->portalExtensionFindAction->find($portalNodeKey);
+            $alreadyActiveExtensions = [];
 
-            foreach ($payload->getPortalExtensionQuery() as $portalExtensionClass) {
-                if (!\class_exists($portalExtensionClass)) {
-                    throw new PortalExtensionMissingException($portalExtensionClass, 1650142325);
+            foreach ($payload->getPortalExtensionQueries() as $query) {
+                $queriedPortalExtensions = $this->packageQueryMatcher->matchPortalExtensions($query, $portalExtensions);
+
+                if ($queriedPortalExtensions->count() === 0) {
+                    throw new NoMatchForPackageQueryException($query, 1650142326);
                 }
 
-                $portalExtensions = $portalExtensions->bySupport($portalNodeGetResult->getPortalClass());
+                foreach ($queriedPortalExtensions as $portalExtension) {
+                    $portalExtensionClass = \get_class($portalExtension);
 
-                if ($portalExtensions->count() === 0) {
-                    throw new PortalExtensionDoesNotSupportPortalNodeException($portalNodeKey, $portalExtensionClass, 1650142326);
-                }
-
-                $portalExtensions = $this->packageQueryMatcher->matchPortalExtensions($portalExtensionClass, $portalExtensions);
-
-                foreach ($portalExtensions as $portalExtension) {
                     if ($portalExtensionState->isActive($portalExtension)) {
-                        throw new PortalExtensionIsAlreadyActiveOnPortalNodeException($portalNodeKey, $portalExtensionClass, 1650142327);
+                        $alreadyActiveExtensions[] = $portalExtensionClass;
+                    } else {
+                        $portalExtensionActivePayload->addExtension($portalExtensionClass);
                     }
                 }
-
-                $portalExtensionActivePayload->addExtension($portalExtensionClass);
             }
 
             if ($portalExtensionActivePayload->getExtensions() === []) {
-                return;
+                throw new PortalExtensionsAreAlreadyActiveOnPortalNodeException($portalNodeKey, $alreadyActiveExtensions, 1650142327);
             }
 
             $this->portalExtensionActivateAction->activate($portalExtensionActivePayload);